PLAT-11060: Reauthenticate inside datafeed loop
If the session expires in the datafeed loop, authentication is going to
be attempted again.

This is only applied to the DF loop, not globally (this is not actively
maintained, doing it globally would require more work). This way the DF
loop should not exit once the session expires.

In case the bot account has been disabled it will keep retrying forever
(again better handling it requires much more work because the authenticate()
part is not propagating errors). Ideally exponential backoff should be
implemented.